requirements.txt
ansi2html                 1.9.1           
asttokens                 2.0.5           
babel                     2.11.0          
blas                      1.0             
blinker                   1.6.2           
bottleneck                1.4.2           
brotli                    1.0.9           
brotli-bin                1.0.9           
brotli-python             1.0.9           
bzip2                     1.0.8           
ca-certificates           2024.11.26      
certifi                   2024.12.14      
charset-normalizer        3.3.2           
click                     8.1.7           
colorama                  0.4.6           
colorlover                0.3.0           
comm                      0.2.1           
contourpy                 1.3.1           
cycler                    0.11.0          
dash                      2.14.2          
dash-bootstrap-components 1.6.0           
debugpy                   1.6.7           
decorator                 5.1.1           
exceptiongroup            1.2.0           
executing                 0.8.3  
gunicorn                  23.0.0
flask                     3.0.3           
flask-compress            1.13            
fonttools                 4.51.0          
freetype                  2.12.1          
idna                      3.7             
importlib-metadata        8.5.0           
intel-openmp              2023.1.0        
ipykernel                 6.29.5          
ipython                   8.30.0          
itsdangerous              2.2.0           
jedi                      0.19.2          
jinja2                    3.1.4           
joblib                    1.4.2           
jpeg                      9e              
jupyter_client            8.6.0           
jupyter_core              5.7.2           
kiwisolver                1.4.4           
lcms2                     2.16            
lerc                      4.0.0           
libbrotlicommon           1.0.9           
libbrotlidec              1.0.9           
libbrotlienc              1.0.9           
libdeflate                1.22            
libffi                    3.4.4           
libpng                    1.6.39          
libsodium                 1.0.18          
libtiff                   4.5.1           
libwebp-base              1.3.2           
lz4-c                     1.9.4           
markupsafe                2.1.3           
matplotlib-base           3.9.2           
matplotlib-inline         0.1.6           
mkl                       2023.1.0        
mkl-service               2.4.0           
mkl_fft                   1.3.11          
mkl_random                1.2.8           
nest-asyncio              1.6.0           
nltk                      3.9.1           
numexpr                   2.10.1          
numpy                     1.26.4          
numpy-base                1.26.4          
openjpeg                  2.5.2           
openssl                   3.0.15          
packaging                 24.2            
pandas                    2.2.3           
parso                     0.8.4           
pillow                    11.0.0          
pip                       24.2            
platformdirs              3.10.0          
plotly                    5.24.1          
prompt-toolkit            3.0.43          
prompt_toolkit            3.0.43          
psutil                    5.9.0           
pure_eval                 0.2.2           
pygments                  2.15.1          
pyparsing                 3.2.0           
pysocks                   1.7.1           
python                    3.10.16         
python-dateutil           2.9.0post0      
python-tzdata             2023.3          
pytz                      2024.1          
pywin32                   305             
pyzmq                     26.2.0          
regex                     2024.9.11       
requests                  2.32.3          
retrying                  1.3.3           
setuptools                75.1.0          
six                       1.16.0          
sqlite                    3.45.3          
stack_data                0.2.0           
tbb                       2021.8.0        
tenacity                  9.0.0           
tk                        8.6.14          
tornado                   6.4.2           
tqdm                      4.66.5          
traitlets                 5.14.3          
typing_extensions         4.12.2          
tzdata                    2024b           
unicodedata2              15.1.0          
urllib3                   2.2.3           
vc                        14.40           
vs2015_runtime            14.42.34433     
wcwidth                   0.2.5           
werkzeug                  3.0.6           
wheel                     0.44.0          
win_inet_pton             1.1.0           
wordcloud                 1.9.3           
xz                        5.4.6           
zeromq                    4.3.5           
zipp                      3.21.0          
zlib                      1.2.13          
zstd                      1.5.6           
